Funqy

Description

The core language operations.

QBit manipulation

new :: Bit -> QM QBit

Create new QBit from a bit. maps \(0 \mapsto |0>\) and \(1 \mapsto |1>\)

measure :: QBit -> QM Bit

Performs a measurement operation, collapsing a QBit to a Bit. Here comes a long explanation of the algorithm correcting the state

Control functions

controlbit :: QM () -> Bit -> QM ()

Sets a classical bit as the controlbit for a quantum gate. Making it run only when the classical bit is equal to one.