QData

Description

This is the core module of the language. This module contains the definitions of all the types exposed to the user.

Core types

newtype QBit (n :: Nat)

Vector state representation of qubit state. Dependent on the number of bits n where the vector becomes \[ \otimes_{i=0}^{n-1} \mathbb{C}^2 \]. The QBit type wraps a statically sized complex vector

Constructors

Q

Fields

getState :: QState (2 ^ n)

data Bit (n :: Nat) where

Type indexed bit strings. Should behave like a list of bits.

Constructors

(:+) :: Bit -> Bit n -> Bit (n + 1) infixr 6

Sing :: Bit -> Bit 1

data Gate (n :: Nat)

Matrix gate representation. Also wraps a function acting on the QBit type

Constructors

Gate

Fields

matrix :: M (2 ^ n) (2 ^ n)

run :: QBit n -> QBit n

type T = ()

The unit type \(* : \top\)

Product type

type family (p :: b) >< (q :: b) :: b infixl 7

The product type family. Represents all types Nat -> * that has a product operation, producing the sum of their type indexed size.

class Prod (p :: Nat -> *) where

Class Prod defines the product operation on sized types

Methods

(><) :: (KnownNat n, KnownNat m, (p n >< p m) ~ p (n + m)) => p n -> p m -> p n >< p m infixl 7

Helpers

bits :: (Storable a, Bits a) => a -> [Bit]

fromMatrix :: KnownNat n => M (2 ^ n) (2 ^ n) -> Gate n

Converts a unitary matrix to the gate type

i :: Complex Double

The imaginary unit