Syntax

Description

Contains all of the language as of now, but will be split in the future

Documentation

type QState = Vector C

The type of the quantum state. \(Q\) in \(\left[Q, L^*, \Lambda \right]\).

data Bit

Bit constructors, largely useless. Might define as a restricted Int instead.

Constructors

Z

O

data Bits

The size that QBit depends on.

Constructors

N Nat

Bits :+ Bits

newtype QBit (n :: Bits)

Vector state representation of qubit state. Dependent on the number of bits n where the vector becomes \( \otimes_{i=0}^{n-1} \mathbb{C}^2 \)

Constructors

Q

Fields

getState :: QState

data ProgramState

Highyl experimental inner type for some stateful monad

Constructors

ProgramState

Fields

vector :: QState

lambda :: *

linker :: *

data m >< n infixr 0

The product type \(\otimes\)

Constructors

m :>< n

new :: Bit -> QBit (N 1)

Constructs new qubits

measure :: QBit (N 1) -> Bit

Collapses a qubit state (of size 1) to a single bit

tensorP :: QBit n -> QBit m -> QBit (n :+ m)

We define the tensor product of qbits as the flattened outer product of the state vectors. Note the summed qubit size.

hadamard :: QBit (N 1) -> QBit (N 1)

Hadamard gate acting on single qubits

cnot :: QBit (N 2) -> QBit (N 2)

CNOT gate acting on two qubits