Syntax

Description

Contains all of the language as of now, but will be split in the future

Documentation

type QState = Vector R

The type of the quantum state. \(Q\) in \(\left[Q, L^*, \Lambda \right]\).

newtype QBit (n :: Nat)

Vector state representation of qubit state. Dependent on the number of bits n where the vector becomes \( \otimes_{i=0}^{n-1} \mathbb{C}^2 \)

Constructors

Q

Fields

getState :: QState

data ProgramState

Highly experimental inner type for some stateful monad

Constructors

ProgramState

Fields

vector :: QState

lambda :: *

linker :: *

data m >< n infixr 0

The product type \(\otimes\)

Constructors

m :>< n

type T = ()

The unit type \(* : \top\)

new :: Bit -> QBit 1

Constructs new qubits

measure :: QBit 1 -> Bit

Collapses a qubit state (of size 1) to a single bit

tensorProduct :: QBit n -> QBit m -> QBit (n + m)

We define the tensor product of qbits as the flattened outer product of the state vectors. Note the summed qubit size.

(><) :: QBit n -> QBit m -> QBit (n + m) infixr 0

Infix synonym for tensorProduct

(|><|) :: Matrix R -> Matrix R -> Matrix R

Kronecker product. Used to run gates in parallel.

gate :: Matrix R -> QBit n -> QBit n

Create quantum gate from its matrix representation

hadamard :: QBit 1 -> QBit 1

Hadamard gate acting on single qubits

cnot :: QBit 2 -> QBit 2

CNOT gate acting on two qubits